feat(vehicle mapping): use nanoseconds and TimeFieldAdapter in vehicle configs
 * All time values in vehicle and vehicle flow generations are now treated in nanoseconds and are stored as longs instead of doubles.
* has the consequence that all mapping_config.json needed to be adjusted
* some cleanup

+++ fed/mosaic-mapping/src/main/java/org/eclipse/mosaic/fed/mapping/ambassador/OriginDestinationVehicleFlowGenerator.java
@@ -38,8 +38,8 @@ public class OriginDestinationVehicleFlowGenerator {
     private final List<List<Double>> odValues;
     private final VehicleDeparture.LaneSelectionMode laneSelectionMode;
     private final VehicleDeparture.DepartureSpeedMode departureSpeedMode;
-    private final double startingTime;
-    private final Double maxTime;
+    private final long startingTime;
+    private final Long maxTime;
 
     /**
      * Constructor for {@link OriginDestinationVehicleFlowGenerator}.
